+++ app/routes/__root.tsx
@@ -1,0 +1,53 @@
+import type { ReactNode } from 'react';
+import {
+  Outlet,
+  createRootRoute,
+  HeadContent,
+  Scripts,
+} from '@tanstack/react-router';
+import indexCss from '../../index.css?url';
+
+export const Route = createRootRoute({
+  head: () => ({
+    meta: [
+      {
+        charSet: 'utf-8',
+      },
+      {
+        name: 'viewport',
+        content: 'width=device-width, initial-scale=1',
+      },
+      {
+        title: 'Combat Tracker',
+      },
+      {
+        rel: 'stylesheet',
+        type: 'text/css',
+        href: indexCss,
+      },
+    ],
+  }),
+  component: RootComponent,
+});
+
+function RootComponent() {
+  return (
+    <RootDocument>
+      <Outlet />
+    </RootDocument>
+  );
+}
+
+function RootDocument({ children }: Readonly<{ children: ReactNode }>) {
+  return (
+    <html>
+      <head>
+        <HeadContent />
+      </head>
+      <body className="size-full h-screen bg-neutral-100">
+        {children}
+        <Scripts />
+      </body>
+    </html>
+  );
+}
