+++ app/ssr.tsx
@@ -1,0 +1,12 @@
+import {
+  createStartHandler,
+  defaultStreamHandler,
+} from '@tanstack/react-start/server';
+import { getRouterManifest } from '@tanstack/react-start/router-manifest';
+
+import { createRouter } from './router';
+
+export default createStartHandler({
+  createRouter,
+  getRouterManifest,
+})(defaultStreamHandler);
