--- src/index.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
-import "./index.css";
-
-const root = ReactDOM.createRoot(
-    document.getElementById("root") as HTMLElement
-);
-root.render(
-    <React.StrictMode>
-        <div className="size-full h-screen bg-neutral-100">
-            <App />
-        </div>
-    </React.StrictMode>
-);
