+++ tsconfig.json
@@ -1,26 +1,43 @@
 {
   "compilerOptions": {
-    "target": "es5",
+    "target": "ES2022",
     "lib": [
       "dom",
       "dom.iterable",
       "esnext"
     ],
+    "baseUrl": ".",
     "allowJs": true,
     "skipLibCheck": true,
+    "strictNullChecks": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
     "module": "esnext",
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx"
+    "jsx": "react-jsx",
+    "paths": {
+        "@components/*": [
+            "Components/*",
+        ],
+        "@ui/*": [
+            "Components/UI/*",
+        ],
+        "@store": [
+            "useStateStore"
+        ],
+        "@types": [
+            "types"
+        ]
+    }
   },
   "include": [
-    "src"
-  ]
+    "app"
+  ],
+  "exclude": ["node_modules", ".vinxi"]
 }
